requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
boto3==1.26.131
botocore==1.29.131
celery==5.2.7
certifi==2023.5.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
crispy-bootstrap5==0.7
cron-descriptor==1.3.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20230430
Django==4.2.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-debug-toolbar==4.0.0
django-import-export==3.2.0
django-redis==5.2.0
django-storages==1.13.2
django-timezone-field==5.0
et-xmlfile==1.1.0
flower==1.2.0
gunicorn==20.1.0
humanize==4.6.0
jmespath==1.0.1
kombu==5.2.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
Pillow==9.5.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
redis==4.5.5
s3transfer==0.6.1
sentry-sdk==1.22.2
six==1.16.0
sqlparse==0.4.4
tablib==3.4.0
tornado==6.3.1
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
xlrd==2.0.1
xlwt==1.3.0
